# swagger.yaml
basePath: /api
definitions:
  database.CardConfirmation:
    properties:
      address:
        type: string
      currency_id:
        type: integer
      id:
        type: integer
      image:
        items:
          type: integer
        type: array
      user_id:
        type: integer
      verified:
        type: boolean
    type: object
  database.CreateCurrencyParams:
    properties:
      code:
        type: string
      description:
        type: string
    type: object
...
  server.ApproveCardConfirmationRequest:
    properties:
      confirmation_id:
        type: integer
    type: object
  server.Balance:
    properties:
      address:
        type: string
      balance:
        type: number
      code:
        type: string
      description:
        type: string
      id:
        type: integer
    type: object
  server.Balances:
    properties:
      balances:
        items:
          $ref: '#/definitions/server.Balance'
        type: array
    type: object
  server.Busy:
    properties:
      busy:
        type: boolean
    type: object
  server.CancelCardRequest:
    properties:
      confirmation_id:
        type: integer
    type: object
  server.CancelOrderRequest:
    properties:
      order_id:
        type: integer
    type: object
  server.CardConfirmationsResponse:
    properties:
      confirmations:
        items:
          $ref: '#/definitions/database.CardConfirmation'
        type: array
    type: object
  server.CreateBalanceRequest:
    properties:
      address:
        type: string
      balance:
        type: number
      currency_id:
        type: integer
    type: object
  server.CreateExchangerRequest:
    properties:
      description:
        type: string
      in_currency:
        type: string
      inmin:
        type: number
      out_currency:
        type: string
      payment_verification:
        type: boolean
    type: object
  server.CreateOrderRequest:
    properties:
      address:
        type: string
      amount:
        type: number
      email:
        type: string
      in_currency:
        type: string
      out_currency:
        type: string
      payment_address:
        type: string
    type: object
  server.CreateOrderResponse:
    properties:
      in_amount:
        type: number
      order_number:
        type: integer
      out_amount:
        type: number
      transfer_address:
        type: string
    type: object
  server.CreateUserRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
...
  server.CurrentRateResponse:
    properties:
      amount:
        type: number
    type: object
...
  server.ExecuteOrderRequest:
    properties:
      order_id:
        type: integer
    type: object
  server.GetCardConfirmationsResponse:
    properties:
      card_confirmations:
        items:
          $ref: '#/definitions/database.CardConfirmation'
        type: array
    type: object
  server.Order:
    properties:
      address:
        type: string
      amountin:
        type: number
      amountout:
        type: number
      approvepic:
        items:
          type: integer
        type: array
      currin:
        type: string
      currout:
        type: string
      email:
        type: string
      id:
        type: integer
      status:
        type: string
    type: object
  server.OrderStatusResponse:
    properties:
      status:
        type: string
    type: object
  server.Orders:
    properties:
      orders:
        items:
          $ref: '#/definitions/server.Order'
        type: array
    type: object
  server.RemoveBalanceRequest:
    properties:
      id:
        type: integer
    type: object
  server.RemoveCurrencyRequest:
    properties:
      code:
        type: string
    type: object
  server.RemoveExchangerRequest:
    properties:
      id:
        type: integer
    type: object
  server.UpdateBalanceRequest:
    properties:
      address:
        type: string
      balance:
        type: number
      balance_id:
        type: integer
      currency_code:
        type: string
    type: object
  server.UserOrder:
    properties:
      id:
        type: integer
      in_amount:
        type: number
      in_currency:
        type: string
      out_addr:
        type: string
      out_amount:
        type: number
      out_currency:
        type: string
      recv_addr:
        type: string
      status:
        type: string
    type: object
  server.UserOrdersResponse:
    properties:
      orders:
        items:
          $ref: '#/definitions/server.UserOrder'
        type: array
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: Swap Support
    url: http://github.com/d1nch8g/swap
  description: Swap exchange api.
  license:
    name: MIT
    url: https://github.com/d1nch8g/swap/src/branch/main/LICENSE
  termsOfService: http://github.com/d1nch8g/swap
  title: Swap exchange api
  version: "1.0"
paths:
  /admin/cancel-order:
    post:
      parameters:
      - description: Cancel order parameters
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/server.CancelOrderRequest'
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Cancel user order and mark it as cancelled
  /admin/change-busy:
    post:
      parameters:
      - description: Busy status
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/server.Busy'
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Change busy status for admin operator
  /admin/check-if-admin:
    delete:
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Check if user is an admin
  /admin/create-balance:
    post:
      parameters:
      - description: Create balance parameters
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/server.CreateBalanceRequest'
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Create new operator balance
  /admin/create-currency:
    post:
      parameters:
      - description: Create currency params
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/database.CreateCurrencyParams'
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Create new currency in exchanger
  /admin/create-exchanger:
    post:
      parameters:
      - description: Create exchanger parameters
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/server.CreateExchangerRequest'
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Create new exchanger with provided currencies
  /admin/execute-order:
    post:
      parameters:
      - description: Execute order parameters
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/server.ExecuteOrderRequest'
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Execute order and change operator balances, update busy
  /admin/list-balances:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.Balances'
      security:
      - ApiKeyAuth: []
      summary: List operator currency balances
  /admin/remove-balance:
    delete:
      parameters:
      - description: Balance id
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/server.RemoveBalanceRequest'
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Remove operators balance
  /admin/remove-currency:
    delete:
      parameters:
      - description: Remove currency parameter
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/server.RemoveCurrencyRequest'
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Remove currency from currency list
  /admin/remove-exchanger:
    delete:
      parameters:
      - description: Remove exchanger parameters
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/server.RemoveExchangerRequest'
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Remove existing exchanger from API
  /confirm-payment:
    post:
      parameters:
      - description: Order id
        in: query
        name: order_id
        required: true
        type: string
      - description: File with operation check
        in: formData
        name: file
        required: true
        type: file
      responses:
        "200":
          description: OK
      summary: Confirm that payment operation is approved and provide check
  /create-order:
    post:
      parameters:
      - description: Request parameters
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/server.CreateOrderRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.CreateOrderResponse'
      summary: Create order to exchange specific currency
  /create-user:
    post:
      parameters:
      - description: Create user request
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/server.CreateUserRequest'
      responses:
        "200":
          description: OK
      summary: Create new user request
  /current-rate:
    get:
      parameters:
      - description: Currency in
        in: path
        name: currency_in
        required: true
        type: string
      - description: Currency out
        in: path
        name: currency_out
        required: true
        type: string
      - description: Amount in
        in: path
        name: amount
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.CurrentRateResponse'
      summary: Current rate at specific currency
...
  /login:
    post:
      parameters:
      - description: Email
        in: header
        name: email
        required: true
        type: string
      - description: Password
        in: header
        name: password
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Login and get auth key
  /operator/approve-card:
    post:
      parameters:
      - description: Approve card request
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/server.ApproveCardConfirmationRequest'
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Mark user credit card as approved
  /operator/cancel-card:
    delete:
      parameters:
      - description: Remove confirmation id
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/server.CancelCardRequest'
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Remove user's card request
  /operator/card-confirmations:
    get:
      parameters:
      - description: Email
        in: query
        name: email
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.CardConfirmationsResponse'
      security:
      - ApiKeyAuth: []
      summary: Get user's card confirmations
  /operator/finished-orders:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.Orders'
      security:
      - ApiKeyAuth: []
      summary: Get finished orders bound to specific operator
  /operator/get-card-confirmations:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.GetCardConfirmationsResponse'
      security:
      - ApiKeyAuth: []
      summary: Get user credit card approval images with parameters
  /operator/get-orders:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.Orders'
      security:
      - ApiKeyAuth: []
      summary: Get active orders bound to specific operator
  /operator/update-balance:
    post:
      parameters:
      - description: Update balance parameters
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/server.UpdateBalanceRequest'
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Update operator currency balance
  /order-status:
    get:
      parameters:
      - description: orderid
        in: header
        name: Orderid
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.OrderStatusResponse'
      summary: Check order status and return info about order
  /user/list-orders:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.UserOrdersResponse'
      security:
      - ApiKeyAuth: []
      summary: List user's orders
  /validate-card:
    post:
      parameters:
      - description: Email
        in: query
        name: email
        required: true
        type: string
      - description: Currency
        in: query
        name: currency
        required: true
        type: string
      - description: Address
        in: query
        name: addr
        required: true
        type: string
      - description: Approve file
        in: formData
        name: file
        required: true
        type: file
      responses:
        "200":
          description: OK
      summary: Create order to exchange specific currency
  /verify/{uuid}:
    get:
      parameters:
      - description: UUID sent by email
        in: path
        name: uuid
        required: true
        type: string
      responses:
        "200":
          description: OK
      summary: Verify user email address
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    description: Token authorization for internal operations
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"